Modifications to tray and startup apps.
Removed nm-applet and a related floating window as I don't think it's
worth it to have a GUI network manager interface when I could just use
nmcli or even nmtui. Made the workspace 1 JSON space for Slack more
specific to grab only the main slack window and not the floating startup
dialog window. So I added slack to the startup apps.