Added custom blocklet for pacman updates.
It first checks for pacman updates, and if any are found displays the
number and sets the color to red. It then checks for AUR updates, and
displays the count as yellow if any updates are found. If no updates are
found it displays as green.

When the pacman blocklet is clicked, octopi is opened.