Adjusted load blocklet for different systems.
Modified the load blocklet so that instance can now be the cpu or thread
count. This way the urgent percent can be relevant depending on the
system. My system has 8 CPUs and the default urgent threashold of 0.9
was much too low. Since instance no longer denotes which load average to
display, the blocklet now only shows the 5 minutes load average. This is
fine for me since it would be the only one I would need for a simplified
load indicator.