Included used scripts.
This are really located in my user's bin directory. The custom lock
pixelates all screens and places a lock icon on the primary screen. This
works on my three monitor setup that uses a thunderbolt DisplayLink dock
and XRandR.

When the window criteria script is ran, it lets you click a window and
is given the i3 criteria of that said window. This has been an
invaluable script for the i3 config and my layout save/restore.